+++ task3/src/Math/Vector/Vector2.java
@@ -1,18 +1,18 @@
 package Math.Vector;
 
 public class Vector2 {
-    private double x, y;
+    private float x, y;
 
-    public Vector2(double x, double y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
@@ -24,30 +24,30 @@ public class Vector2 {
         return new Vector2(this.x - other.x, this.y - other.y);
     }
 
-    public Vector2 multiply(double scalar) {
+    public Vector2 multiply(float scalar) {
         return new Vector2(this.x * scalar, this.y * scalar);
     }
 
-    public Vector2 divide(double scalar) {
+    public Vector2 divide(float scalar) {
         if (scalar == 0) {
             throw new IllegalArgumentException("Мы не делим на 0");
         }
         return new Vector2(this.x / scalar, this.y / scalar);
     }
 
-    public double length() {
-        return Math.sqrt(x * x + y * y);
+    public float length() {
+        return (float) Math.sqrt(x * x + y * y);
     }
 
     public Vector2 normalize() {
-        double length = length();
+        float length = length();
         if (length == 0) {
             throw new ArithmeticException("Ты точно хочешь нормализовать 0-вектор?");
         }
         return this.divide(length);
     }
 
-    public double dotProduct(Vector2 other) {
+    public float dotProduct(Vector2 other) {
         return this.x * other.x + this.y * other.y;
     }
 }
