+++ task3/src/Test/TestMatrix3x3.java
@@ -10,12 +10,12 @@ import static org.junit.Assert.assertEquals;
 public class TestMatrix3x3 {
     @Test
     public void testPlus() {
-        double[][] data1 = {
+         float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+         float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -25,7 +25,7 @@ public class TestMatrix3x3 {
 
         Matrix3x3 result = m1.plus(m2);
 
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {10, 10, 10},
                 {10, 10, 10},
                 {10, 10, 10}
@@ -40,12 +40,12 @@ public class TestMatrix3x3 {
 
     @Test
     public void testMinus() {
-        double[][] data1 = {
+         float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+         float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -55,7 +55,7 @@ public class TestMatrix3x3 {
 
         Matrix3x3 result = m1.minus(m2);
 
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {-8, -6, -4},
                 {-2, 0, 2},
                 {4, 6, 8}
@@ -71,7 +71,7 @@ public class TestMatrix3x3 {
 
     @Test
     public void testMultiplyByVector() {
-        double[][] data = {
+         float[][] data = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
@@ -89,12 +89,12 @@ public class TestMatrix3x3 {
 
     @Test
     public void testMultiply() {
-        double[][] data1 = {
+         float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+         float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -102,7 +102,7 @@ public class TestMatrix3x3 {
         Matrix3x3 m1 = new Matrix3x3(data1);
         Matrix3x3 m2 = new Matrix3x3(data2);
         Matrix3x3 result = m1.multiply(m2);
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {30, 24, 18},
                 {84, 69, 54},
                 {138, 114, 90}
@@ -117,14 +117,14 @@ public class TestMatrix3x3 {
 
     @Test
     public void testTranspose() {
-        double[][] data = {
+         float[][] data = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
         Matrix3x3 matrix = new Matrix3x3(data);
         Matrix3x3 result = matrix.transpose();
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {1, 4, 7},
                 {2, 5, 8},
                 {3, 6, 9}
@@ -140,36 +140,36 @@ public class TestMatrix3x3 {
     @Test
     public void testIdentity() {
         Matrix3x3 identity = Matrix3x3.identity();
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {1, 0, 0},
                 {0, 1, 0},
                 {0, 0, 1}
         };
-        double[][] actualData = identity.getElements();
+        float[][] actualData = identity.getElements();
         assertArrayEquals(expectedData, actualData);
     }
 
     @Test
     public void testZeroMatrix() {
         Matrix3x3 zero = Matrix3x3.zeroMatrix();
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {0, 0, 0},
                 {0, 0, 0},
                 {0, 0, 0}
         };
-        double[][] actualData = zero.getElements();
+        float[][] actualData = zero.getElements();
         assertArrayEquals(expectedData, actualData);
     }
 
     @Test
     public void testDeterminant(){
-        double[][] data = {
+         float[][] data = {
                 {1, 2, 4},
                 {4, 15, 6},
                 {2, 8, 9}
         };
         Matrix3x3 matrix = new Matrix3x3(data);
-        double result = matrix.determinant();
+         float result = matrix.determinant();
         assertEquals(47.0, result, 0.001);
     }
 }
