+++ task3/src/Test/TestMatrix4x4.java
@@ -12,13 +12,13 @@ public class TestMatrix4x4 {
 
     @Test
     public void testPlus() {
-        double[][] data1 = {
+         float[][] data1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
                 {13, 14, 15, 16}
         };
-        double[][] data2 = {
+         float[][] data2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -27,7 +27,7 @@ public class TestMatrix4x4 {
         Matrix4x4 m1 = new Matrix4x4(data1);
         Matrix4x4 m2 = new Matrix4x4(data2);
         Matrix4x4 result = m1.plus(m2);
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {17, 17, 17, 17},
                 {17, 17, 17, 17},
                 {17, 17, 17, 17},
@@ -38,13 +38,13 @@ public class TestMatrix4x4 {
 
     @Test
     public void testMinus() {
-        double[][] data1 = {
+         float[][] data1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
                 {13, 14, 15, 16}
         };
-        double[][] data2 = {
+         float[][] data2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -53,7 +53,7 @@ public class TestMatrix4x4 {
         Matrix4x4 m1 = new Matrix4x4(data1);
         Matrix4x4 m2 = new Matrix4x4(data2);
         Matrix4x4 result = m1.minus(m2);
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {-15, -13, -11, -9},
                 {-7, -5, -3, -1},
                 {1, 3, 5, 7},
@@ -64,14 +64,14 @@ public class TestMatrix4x4 {
 
     @Test
     public void testMultiplyByVector() {
-        double[][] matrixData = {
+         float[][] matrixData = {
                 {2, 0, 0, 0},
                 {0, 2, 0, 0},
                 {0, 0, 2, 0},
                 {0, 0, 0, 2}
         };
         Matrix4x4 matrix = new Matrix4x4(matrixData);
-        double[] vectorData = {1, 2, 3, 4};
+         float[] vectorData = {1, 2, 3, 4};
         Vector4 vector = new Vector4(vectorData[0], vectorData[1],vectorData[2], vectorData[3]);
         Vector4 result = matrix.multiplyByVector(vector);
         assertEquals(2, result.getX(), 0.01);
@@ -82,14 +82,14 @@ public class TestMatrix4x4 {
 
     @Test
     public void testMultiply() {
-        double[][] matrixData1 = {
+         float[][] matrixData1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
                 {13, 14, 15, 16}
         };
         Matrix4x4 matrix1 = new Matrix4x4(matrixData1);
-        double[][] matrixData2 = {
+         float[][] matrixData2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -97,7 +97,7 @@ public class TestMatrix4x4 {
         };
         Matrix4x4 matrix2 = new Matrix4x4(matrixData2);
         Matrix4x4 result = matrix1.multiply(matrix2);
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {80, 70, 60, 50},
                 {240, 214, 188, 162},
                 {400, 358, 316, 274},
@@ -108,7 +108,7 @@ public class TestMatrix4x4 {
 
     @Test
     public void testTranspose() {
-        double[][] matrixData = {
+         float[][] matrixData = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
@@ -116,7 +116,7 @@ public class TestMatrix4x4 {
         };
         Matrix4x4 matrix = new Matrix4x4(matrixData);
         Matrix4x4 transposed = matrix.transpose();
-        double[][] expectedData = {
+         float[][] expectedData = {
                 {1, 5, 9, 13},
                 {2, 6, 10, 14},
                 {3, 7, 11, 15},
@@ -128,8 +128,8 @@ public class TestMatrix4x4 {
     @Test
     public void testIdentity() {
         Matrix4x4 identity = Matrix4x4.identity();
-        double[][] data = identity.getElements();
-        double[][] expectedData = {
+         float[][] data = identity.getElements();
+         float[][] expectedData = {
                 {1, 0, 0, 0},
                 {0, 1, 0, 0},
                 {0, 0, 1, 0},
@@ -141,8 +141,8 @@ public class TestMatrix4x4 {
     @Test
     public void testZeroMatrix() {
         Matrix4x4 zero = Matrix4x4.zeroMatrix();
-        double[][] data = zero.getElements();
-        double[][] expectedData = {
+         float[][] data = zero.getElements();
+         float[][] expectedData = {
                 {0, 0, 0, 0},
                 {0, 0, 0, 0},
                 {0, 0, 0, 0},
@@ -154,14 +154,14 @@ public class TestMatrix4x4 {
 
     @Test
     public void testDetermination(){
-        double[][] data = {
+         float[][] data = {
                 {1, 7, 1, 8},
                 {4, 6, 11, 8},
                 {1, 3, 1, 8},
                 {2, 5, 9, 1}
         };
         Matrix4x4 matrix = new Matrix4x4(data);
-        double result = matrix.determinant();
+         float result = matrix.determinant();
         assertEquals(252, result, 0.001);
     }
 }
